+++ src/script/page/LeftSidebar/panels/Preferences.test.tsx
